requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.3
django-cors-headers==3.5.0
django-redis==4.12.1
djangorestframework==3.12.2
importlib-metadata==3.1.1
kombu==5.0.2
mysqlclient==2.0.1
prompt-toolkit==3.0.8
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
